Ignore unrecognised socket families in from_host
On MacOS, SIOCGIFCONF returns sockets with a number of different families, including AF_LINK, AF_PUP, AF_LAT, and a few others with codes greater than the highest entry on the list. Rather than adding specific handling for each of these, we assume they are not of interest for this use case and ignore/log them.